Update test for BaseLayout components
