--- src/components/containers/BaseLayout/components/SlideMenu/SlideMenu.test.js
@@ -1,26 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-
-// Enzyme imports
-import Enzyme, { mount, shallow } from 'enzyme';
-import Adapter from 'enzyme-adapter-react-16';
-
-// Component import
-import SlideMenu from './SlideMenu.js';
-import navLinks from 'constants/linksData.js';
-
-Enzyme.configure({ adapter: new Adapter() });
-
-const defaultProps = {
-    navLinks: navLinks,
-};
-
-const testElement = <SlideMenu {...defaultProps} />;
-
-describe('<SlideMenu />', () => {
-
-    it('renders without crashing', () => {
-        const div = document.createElement('div');
-        ReactDOM.render(testElement, div);
-    });
-});
