--- src/components/containers/BaseLayout/components/TopMenu/TopMenu.test.js
@@ -1,27 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import { BrowserRouter } from 'react-router-dom';
-
-// Enzyme imports
-import Enzyme, { mount, shallow } from 'enzyme';
-import Adapter from 'enzyme-adapter-react-16';
-
-// Component import
-import TopMenu from './TopMenu.js';
-
-Enzyme.configure({ adapter: new Adapter() });
-
-const defaultProps = {};
-
-const testElement =
-  <BrowserRouter>
-    <TopMenu {...defaultProps} />
-  </BrowserRouter>;
-
-describe('<TopMenu />', () => {
-
-  it('renders without crashing', () => {
-    const div = document.createElement('div');
-    ReactDOM.render(testElement, div);
-  });
-});
